norme ok, pb de compilations réglés, pret a tester

+++ ft_strlcat.c
@@ -6,16 +6,16 @@
 /*   By: tlevaufr <tlevaufr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/09 19:13:26 by tlevaufr          #+#    #+#             */
-/*   Updated: 2017/11/21 18:16:27 by tlevaufr         ###   ########.fr       */
+/*   Updated: 2017/11/22 16:12:23 by tlevaufr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
